% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.r
\name{as_date}
\alias{as_date}
\title{Shortcut to avoid specifying origin}
\usage{
as_date(x, ...)
}
\arguments{
\item{x}{an object to be converted}

\item{\dots}{further arguments to be passed from or to other methods}
}
\description{
Shortcut to avoid specifying origin
}
