% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half-baked-fstring.r
\name{fstr}
\alias{fstr}
\title{Half-baked "f strings"}
\usage{
fstr(x)
}
\arguments{
\item{x}{string to evaluate}
}
\value{
new string with evaluated replacements for all \code{\{strs\}}.
}
\description{
Replaces \code{"\{string\}"}s embedded in a string with their evaluated values and
returns the string. It's a dangerous, lightweight version of what the \code{whisker}
package does and a less-than-half implementation of Python 3.6.x+ f-strings.
}
\details{
If the evaluation ultimately fails, the \code{\{string\}} is replaces with \code{NA}.

NOTE: THIS IS A DANGEROUS FUNCTION as you'll see in the examples
}
\note{
In case you haven't noticed, this is a very dangerous function.
}
\examples{
i <- 10
n <- 10
name <- "boB"

for (i in 1:n) {
  name <- "Bill"
  print(fstr("This iteration {i} of {n}, {name}. And, this is [{mtcars$wt[1]} and {rnorm(100)[1]} and (DANGEROUSLY) {system('wc -l /etc/passwd', intern=TRUE)}]"))
}
}

