% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbarplot.r
\name{gbarplot}
\alias{gbarplot}
\alias{gbarplot.data.frame}
\alias{gbarplot.list}
\alias{gbarplot.table}
\alias{gbarplot.vector}
\title{Quick way to replace the handy \code{barplot}}
\usage{
gbarplot(x, fill = "maroon", col = "black")

\method{gbarplot}{table}(x, fill = "maroon", col = "black")

\method{gbarplot}{vector}(x, fill = "maroon", col = "black")

\method{gbarplot}{list}(x, fill = "maroon", col = "black")

\method{gbarplot}{data.frame}(x, fill = "maroon", col = "black")
}
\description{
Quick way to replace the handy \code{barplot}
}

