% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{ggplot_with_subtitle}
\alias{ggplot_with_subtitle}
\title{Add a subtitle to a ggplot object and draw plot on current graphics device.}
\usage{
ggplot_with_subtitle(gg, label = "", fontfamily = NULL, fontsize = 10,
  hjust = 0, vjust = 0, bottom_margin = 5.5, newpage = is.null(vp),
  vp = NULL, ...)
}
\arguments{
\item{gg}{ggplot2 object}

\item{label}{subtitle label}

\item{fontfamily}{font family to use. The function doesn't pull any font
information from \code{gg} so you should consider specifying fonts
for the plot itself and here. Or send me code to make this smarter :-)}

\item{fontsize}{font size}

\item{hjust, vjust}{horizontal/vertical justification}

\item{bottom_margin}{space between bottom of subtitle and plot (code{pts})}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{parameters passed to \code{gpar} in call to \code{textGrob}}
}
\value{
Invisibly returns the result of \code{\link{ggplot_build}}, which
  is a list with components that contain the plot itself, the data,
  information about the scales, panels etc.
}
\description{
Add a subtitle to a ggplot object and draw plot on current graphics device.
}

