% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match.R
\name{\%fmin\%}
\alias{\%fmin\%}
\title{Shortcut for fmatch - fastmatch version of \code{in}}
\usage{
x \%fmin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against}
}
\description{
fmatch is a faster version of the built-in \code{match()} function. It is
slightly faster than the built-in version because it uses more
specialized code, but in addition it retains the hash table within the
table object such that it can be re-used, dramatically reducing the look-up
time especially for large tables.
}

