% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match.R
\name{\%notfmin\%}
\alias{\%notfmin\%}
\title{"Not in" (fastmatch version)}
\usage{
x \%notfmin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against}
}
\description{
Avoids awkward-to-type \code{!a \%fmin\% b}
}

