% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match.R
\name{\%notin\%}
\alias{\%notin\%}
\title{"Not in"}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against}
}
\description{
Avoids awkward-to-type \code{!a \%in\% b}
}

