% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/web.r
\name{has_bom}
\alias{has_bom}
\title{Tests whether raw httr response or character vector has a byte-order-mark}
\usage{
has_bom(resp)
}
\arguments{
\item{resp}{\code{httr} response object or a character vector}
}
\value{
\code{TRUE} if response contains a byte-order-mark
}
\description{
Tests whether raw httr response or character vector has a byte-order-mark
}
\examples{
library(httr)
url <- "http://www.claritin.com/webservice/allergyforecast.php?zip=33956"
foo <- GET(url)
has_bom(foo)

cfoo <- content(foo, as="text")
has_bom(cfoo)
}

