% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaches.r
\name{hhs_breaches}
\alias{hhs_breaches}
\title{Retrieve current public list of U.S. Health & Human Services Breaches}
\usage{
hhs_breaches()
}
\description{
Recrods only for those impacting 500 or more persons
}
