% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networking.r
\name{host_up}
\alias{host_up}
\title{Is a host listening on a port (TCP)?}
\usage{
host_up(x, port = 80, timeout = 2, quiet = FALSE)
}
\arguments{
\item{x}{character vector of hosts/ips}

\item{port}{a single port or a vector of the same length as \code{x}; default is port \code{80}}

\item{timeout}{numner of seconds to give the host to respond to the connection request}

\item{quiet}{if \code{FALSE} then a message will be printed for each host contacted}
}
\value{
logical vector; \code{TRUE} == host was reachable on the port within the timeout
        specified
}
\description{
Given a vector of hosts (or IPs) and ports (or a single port)
this function will test whether the node(s) is/are listening for
connections. Fully vectorized with customizable timeouts and options
(on by default) status messages.
}

