% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbrmisc-package.r
\name{hrbrmisc-exports}
\alias{\%>\%}
\alias{hrbrmisc-exports}
\title{hrbrmisc exported operators}
\description{
The following functions are imported and then re-exported
from the hrbrmisc package to enable use of the magrittr
pipe operator with no additional library calls
}

