% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-and-shortcuts.r
\name{is_empty_string}
\alias{is_empty_string}
\title{Is a string equivalent to \code{""}}
\usage{
is_empty_string(x)
}
\arguments{
\item{x}{character vector}
}
\value{
\code{TRUE} if \code{x} == \code{""}
}
\description{
Designed for use as a \code{purr::keep} or \code{purrr::discard}
predicate (or any functional predicate taking a single object as a value).
}
\details{
Works as a standalone vectorized function equally as well.
}
\seealso{
Other purr_predicates: 
\code{\link{is_zero_length}()}
}
\concept{purr_predicates}
