% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-and-shortcuts.r
\name{is_zero_length}
\alias{is_zero_length}
\title{Is length 0}
\usage{
is_zero_length(x)
}
\arguments{
\item{x}{vector}
}
\value{
\code{TRUE} if \code{x} length is \code{0}
}
\description{
Designed for use as a \code{purr::keep} or \code{purrr::discard}
predicate (or any functional predicate taking a single object as a value).
}
\details{
Makes a broad assumption about \code{x} being a singluar object
}
\seealso{
Other purr_predicates: 
\code{\link{is_empty_string}()}
}
\concept{purr_predicates}
