% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonview.r
\name{jsontreeviewOutput}
\alias{jsontreeviewOutput}
\title{Widget output function for use in Shiny}
\usage{
jsontreeviewOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{outputId}

\item{width}{width}

\item{height}{height}
}
\description{
Widget output function for use in Shiny
}

