% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonview.r
\name{jsonviewOutput}
\alias{jsonviewOutput}
\title{Widget output function for use in Shiny}
\usage{
jsonviewOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{outputId}

\item{width}{width}

\item{height}{height}
}
\description{
Widget output function for use in Shiny
}

