% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_numeric}
\alias{make_numeric}
\title{Clean up a character vector to make it numeric}
\usage{
make_numeric(x)
}
\arguments{
\item{x}{character vector to process}
}
\value{
numeric vector
}
\description{
Remove commas & whitespace, primarily
}

