% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_percent}
\alias{make_percent}
\title{Clean up a character vector to make it a percent}
\usage{
make_percent(x)
}
\arguments{
\item{x}{character vector to process}
}
\value{
numeric vector
}
\description{
Remove "%" primarily, convert to numeric & divide by 100
}

