% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.r
\name{map_field_list}
\alias{map_field_list}
\title{data first lapply `[[`}
\usage{
map_field_list(x, field)
}
\arguments{
\item{x}{list/vector}

\item{field}{thing for `[[field]]`}
}
\description{
data first lapply `[[`
}
