% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{\%fmin\%}
\alias{\%!fmin\%}
\alias{\%!in\%}
\alias{\%fmin\%}
\title{Shortcuts for \code{fmatch} and \code{match} operations}
\usage{
x \%fmin\% table

x \%!in\% table

x \%!fmin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against}
}
\description{
\code{fmatch} is a faster version of the built-in \code{match()} function. It is
slightly faster than the built-in version because it uses more
specialized code, but in addition it retains the hash table within the
table object such that it can be re-used, dramatically reducing the look-up
time especially for large tables.
}
\details{
\itemize{
  \item \code{\%notin\%} avoids awkward-to-type \code{!a \%in\% b}
  \item \code{\%fmin\%} fastmatch version of \code{in}
  \item \code{\%notfmin\%} Avoids awkward-to-type \code{!a \%fmin\% b}
}
}

