% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.r
\name{mcga}
\alias{mcga}
\title{Make column names great again}
\usage{
mcga(x)
}
\arguments{
\item{col_names}{character vector}
}
\value{
lower-case, no punct, underscored strings
}
\description{
Make column names great again
}
