% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.r
\name{noathenb}
\alias{\%||\%}
\alias{noathenb}
\title{good for nulls}
\usage{
noathenb(a, b)

a \%||\% b
}
\description{
good way of getting a value if \code{a} is \code{NULL}

good way of getting a value if \code{a} is \code{NULL}
}

