% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-and-shortcuts.r
\name{null-or-zero-or-na}
\alias{null-or-zero-or-na}
\alias{\%||0NA\%}
\title{Is an object \code{NULL} or length zero or \code{NA}}
\usage{
x \%||0NA\% y
}
\arguments{
\item{x}{vector}
}
\value{
\code{TRUE} if \code{x} is \code{NULL} or length \code{0} or \code{NA}
}
\description{
Makes a broad assumption about \code{x} being a singluar object
}
\seealso{
Other shortcuts: 
\code{\link{null-or-zero}}
}
\concept{shortcuts}
