% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates-and-shortcuts.r
\name{null-or-zero}
\alias{\%||0\%}
\alias{null-or-zero}
\title{Is an object \code{NULL} or length zero}
\usage{
x \%||0\% y
}
\arguments{
\item{x}{vector}
}
\value{
\code{TRUE} if \code{x} is \code{NULL} or length \code{0}
}
\description{
Makes a broad assumption about \code{x} being a singluar object
}
\seealso{
Other shortcuts: \code{\link{null-or-zero-or-na}}
}

