% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maps.R
\name{quick_map}
\alias{quick_map}
\title{Plot a fortified map using coord_quickmap}
\usage{
quick_map(map_data, map_id = "region")
}
\arguments{
\item{map_id}{since map data for ggplot is sometimes inconsistent in
the naming scheme for the column providing the \code{map_id},
use "\code{region}" by default but let the caller specify
another name (e.g. "\code{id}") if it's different.}

\item{map_d}{the fortified map data frame}
}
\value{
ggplot object
}
\description{
Plot a fortified map using coord_quickmap
}

