% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.r
\name{sans_bom}
\alias{sans_bom}
\title{Remove byte order mark (BOM) from \code{httr::response} object or character vector}
\usage{
sans_bom(resp)
}
\arguments{
\item{resp}{\code{httr} response object or a character vector}
}
\value{
Converted content
}
\description{
Performs the equivalent of \code{content(resp, type="text")} but removes the
BOM before doing so. If a BOM is not detected, the function will return an error.
}
\details{
\href{https://en.wikipedia.org/wiki/Byte_order_mark}{What is a byte order mark?}
}
\examples{
library(httr)
foo <- GET("http://www.claritin.com/webservice/allergyforecast.php?zip=33956")
sans_bom(foo)
}
\references{
\href{http://www.unicode.org/faq/utf_bom.html}{UTF-8, UTF-16, UTF-32 & BOM}
}
\author{
@hrbrmstr
}
