% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sdbm_hash}
\alias{sdbm_hash}
\title{Personal hash function}
\usage{
sdbm_hash(x)
}
\arguments{
\item{x}{character vector of strings to hash}
}
\value{
numeric vector of the hashes of \code{x}
}
\description{
Personal hash function
}

