% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.r
\name{spectre}
\alias{spectre}
\title{Setup connection to my local phantomjs webdriver}
\usage{
spectre(host = Sys.getenv("PHANTOM_WEBDRIVER_HOST"),
  port = Sys.getenv("PHANTOM_WEBDRIVER_PORT"))
}
\arguments{
\item{host}{host of the phantomjs webdriver instance}

\item{port}{port of the phantomjs webdriver instance}
}
\description{
For scraping efforts it's handy to have a webdriver always going. I happen
to have a spare system (my local mini-CRAN mirror) where I can keep a live
phantomjs node active.
}
\details{
Relies on two environment variables being set since you could theoretically
target the node and exploit webkit vulns if I hardcoded the server info.
}

