% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.r
\name{strsplit_one}
\alias{strsplit_one}
\alias{map_field_simple}
\alias{map_field_list}
\alias{do_call}
\alias{pf_grep}
\alias{pf_grepl}
\alias{pf_sub}
\alias{pf_gsub}
\alias{row_bind}
\alias{row_bind_lst}
\title{Personal tidyverse alternatives}
\usage{
strsplit_one(x, ...)

map_field_simple(x, field)

map_field_list(x, field)

do_call(dat, ƒ)

pf_grep(dat, pat, ...)

pf_grepl(dat, pat, ...)

pf_sub(dat, pat, repl, ...)

pf_gsub(dat, pat, repl, ...)

row_bind(...)

row_bind_lst(lst, ...)
}
\description{
Mostly to get "data" out in front for base piping
}
