% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.r
\name{theme_hrbrmstr_if}
\alias{theme_hrbrmstr_if}
\title{InterFace theme}
\usage{
theme_hrbrmstr_if(...)
}
\description{
InterFace theme
}
\note{
PNG OUTPUT ONLY; The font is horribadly broken from Adobe TypeKit
}

