% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{theme_hrbrmstr_kerkis}
\alias{theme_hrbrmstr_kerkis}
\title{Kerkis theme}
\usage{
theme_hrbrmstr_kerkis(grid = TRUE, axis = FALSE, ticks = FALSE)
}
\description{
"Kerkis (C) Department of Mathematics, University of the Aegean".
}
\details{
\url{http://myria.math.aegean.gr/kerkis/}
}

