% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.r
\name{theme_hrbrmstr_mplus}
\alias{theme_hrbrmstr_mplus}
\title{M+ font theme}
\usage{
theme_hrbrmstr_mplus(base_family = "M+ 1c", base_size = 11,
  strip_text_family = "M+ 1c regular", strip_text_size = 12,
  plot_title_family = "M+ 1c bold", plot_title_size = 18,
  plot_title_margin = 10, plot_margin = margin(30, 30, 30, 30),
  subtitle_family = "M+ 1c medium", subtitle_size = 12,
  subtitle_margin = 15, caption_family = "M+ 1c light", caption_size = 9,
  caption_margin = 10, axis_title_family = subtitle_family,
  axis_title_size = 9, axis_title_just = "rt", grid = TRUE,
  axis = FALSE, ticks = FALSE)
}
\description{
M+ font theme
}
