% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-helpers.r
\name{title_left}
\alias{title_left}
\title{Completely flush-left-align ggplot2 title}
\usage{
title_left(gg)
}
\arguments{
\item{gg}{ggplot2 object}
}
\value{
\code{gtable} object
}
\description{
Completely flush-left-align ggplot2 title
}

