% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep.r
\name{vgrep}
\alias{ggrep}
\alias{ggrepl}
\alias{igrep}
\alias{igrepl}
\alias{lgrep}
\alias{vgrep}
\alias{vigrep}
\title{Shorthand helpers for base \code{grep} and \code{grepl}}
\usage{
vgrep(x, pattern)

ggrep(x, pattern)

igrep(x, pattern)

vigrep(x, pattern)

igrepl(x, pattern)

ggrepl(x, pattern)

lgrep(x, pattern)
}
\arguments{
\item{x}{character vector}

\item{pattern}{regex to look for}
}
\description{
Each puts the object first then the pattern and auto-fills in certain params
}
\details{
\itemize{
  \item \code{ggrep} is the equivalent of \code{grep(pattern, x)}
  \item \code{vgrep} is the equivalent of \code{grep(pattern, x, value=TRUE)}
  \item \code{igrep} is the equivalent of \code{grep(pattern, x, ignore.case=TRUE)}
  \item \code{vigrep} is the equivalent of \code{grep(pattern, x, value=TRUE, ignore.case=TRUE)}
  \item \code{ggrepl} is the equivalent of \code{grepl(pattern, x)}
  \item \code{igrepl} is the equivalent of \code{grepl(pattern, x, ignore.case=TRUE)}
  \item \code{lgrep} is the equivalent of \code{grepl(pattern, x)}
}
}

