% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maps.R
\name{quick_map}
\alias{quick_map}
\title{Plot a fortified map using coord_quickmap}
\usage{
quick_map(map_data, map_id = NULL)
}
\arguments{
\item{map_id}{since map data for ggplot is sometimes inconsistent in
the naming scheme for the column providing the \code{map_id},
this function will try to determine if "\code{region}" or "\code{id}"
is being used but will use the caller-specified \code{map_id} if
present.}

\item{map_d}{the fortified map data frame}
}
\value{
ggplot object
}
\description{
This function will try to automatically determine what column in
\code{map_data} should be used for the \code{map_id}. Specifying a
value overrides this heuristic.
}

