% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.r
\name{sGET}
\alias{sGET}
\alias{sHEAD}
\alias{sPOST}
\title{"Safe" versions of \code{httr} functions}
\usage{
sGET(...)

sPOST(...)

sHEAD(...)
}
\arguments{
\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link{modify_url}}. Unnamed parameters will be combined
with \code{\link{config}}.}
}
\description{
"Safe" versions of \code{httr} functions
}

