% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr.r
\name{sGET}
\alias{sDELETE}
\alias{sGET}
\alias{sHEAD}
\alias{sPATCH}
\alias{sPOST}
\alias{sPUT}
\alias{sVERB}
\title{"Safe" versions of \code{httr} functions}
\usage{
sGET(...)

sPOST(...)

sHEAD(...)

sPATCH(...)

sPUT(...)

sDELETE(...)

sVERB(...)
}
\arguments{
\item{\dots}{Passed through to the \code{httr} verb function.}
}
\description{
Each \code{httr} verb is wrapped with \code{purrr::safely()}
}

