% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{theme_hrbrmstr}
\alias{theme_hrbrmstr}
\title{My ggplot2 go-to theme}
\usage{
theme_hrbrmstr(base_family = "OpenSans-CondensedLight", base_size = 11,
  strip_text_family = base_family, strip_text_size = 12,
  plot_title_family = "OpenSans-CondensedBold", plot_title_size = 18,
  plot_title_margin = 16,
  axis_title_family = "OpenSans-CondensedLightItalic", axis_title_size = 9,
  axis_title_just = "rt", grid = TRUE, axis = FALSE, ticks = FALSE)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{strip_text_family}{facet label font family}

\item{strip_text_size}{facet label text size}

\item{plot_title_family}{plot tilte family}

\item{plot_title_size}{plot title font size}

\item{plot_title_margin}{plot title margin}

\item{axis_title_family}{axis title font family}

\item{axis_title_size}{axis title font size}

\item{axis_title_just}{axis title font justification \code{blmcrt}}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of
\code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{axis}

\item{ticks}{ticks}
}
\description{
It requires installing Open Sans Condensed fonts from Google Fonts
unless you change the font parameters
}

