% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.r
\name{theme_hrbrmstr}
\alias{theme_hrbrmstr}
\title{My ggplot2 go-to theme}
\usage{
theme_hrbrmstr(base_family = "Cabin-Regular", base_size = 11,
  strip_text_family = base_family, strip_text_size = 12,
  plot_title_family = "Cabin-Bold", plot_title_size = 18,
  plot_title_margin = 10, plot_margin = margin(10, 10, 10, 10),
  subtitle_family = "Cabin-Italic", subtitle_size = 12,
  subtitle_margin = 15, caption_family = "Cabin-Regular",
  caption_size = 9, caption_margin = 10,
  axis_title_family = subtitle_family, axis_title_size = 9,
  axis_title_just = "rt", grid = TRUE, axis = FALSE, ticks = FALSE)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{strip_text_family}{facet label font family}

\item{strip_text_size}{facet label text size}

\item{plot_title_family}{plot tilte family}

\item{plot_title_size}{plot title font size}

\item{plot_title_margin}{plot title margin}

\item{plot_margin}{plot margin}

\item{subtitle_family}{plot subtitle family}

\item{subtitle_size}{plot subtitle size}

\item{subtitle_margin}{plot subtitle margin}

\item{caption_family}{plot caption family}

\item{caption_size}{plot caption size}

\item{caption_margin}{plot caption margin}

\item{axis_title_family}{axis title font family}

\item{axis_title_size}{axis title font size}

\item{axis_title_just}{axis title font justification \code{blmcrt}}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of
\code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{axis \code{TRUE}, \code{FALSE}, [\code{xy}]}

\item{ticks}{ticks}
}
\description{
It requires installing Cabin fonts unless you change the font parameters
}
\details{
\url{http://www.impallari.com/cabin/}
}

